+++ src/ShellUtils.php
@@ -20,7 +20,6 @@
 namespace Kristuff\AbuseIPDB;
 
 use Kristuff\Mishell\Console;
-use Kristuff\Mishell\Program;
 use Kristuff\AbuseIPDB\ApiHandler;
 
 /**
@@ -33,7 +32,7 @@ abstract class ShellUtils
     /**
      * helper functions
      */
-    use UtilsTrait, ErrorsTrait;
+    use UtilsTrait;
   
     const OUTPUT_JSON       = 'json';
     const OUTPUT_DEFAULT    = 'default';
@@ -253,160 +252,4 @@ abstract class ShellUtils
         $badge = str_pad($score, 3, ' ',STR_PAD_LEFT); 
         return Console::text($padding.$badge.$padding, $scoreforegroundColor, $scoreBackgroundColor);
     }
-
-    /**
-     * Check and print errors in API response. 
-     * 
-     * @access protected
-     * @static
-     * @param object     $response       
-     * @param bool       $checkForEmpty     
-     * 
-     * @return bool     
-     */
-    protected static function hasErrors(object $response, bool $checkForEmpty = true )
-    {
-        return $checkForEmpty ? self::parseErrors($response) || self::checkForEmpty($response) : self::parseErrors($response);
-    }
-
-    /**
-     * Check and print errors in API response. 
-     * 
-     * @access protected
-     * @static
-     * @param object     $response       
-     * @param bool       $checkForEmpty     
-     * 
-     * @return bool     
-     */
-    private static function parseErrors(object $response)
-    {
-        if (isset($response) && isset($response->errors)){
-            switch (self::$outputFormat){
-                case self::OUTPUT_DEFAULT:
-                    self::printFormattedErrors($response);
-                    break;
-
-                case self::OUTPUT_PLAINTEXT:
-                    self::printPlainTextErrors($response);
-                    break;
-
-                case self::OUTPUT_JSON:
-                    echo json_encode($response, JSON_PRETTY_PRINT);
-                    break;
-            }
-            return true;
-        }
-        return false;    
-    }
-
-    /**
-     * 
-     * @access protected
-     * @static
-     * @param object     $response       
-     * 
-     * @return void     
-     */
-    protected static function printFormattedErrors(object $response)
-    {
-        // top error badge    
-        Console::log('  ' .   Console::text(' ERROR ','white', 'red'));
-
-        $num = 0;
-        // errors is an array, could have more than one error..
-        foreach ($response->errors as $err){
-            $num++;
-
-            Console::log(Console::text('   ✗', 'red') .  self::printResult(' Number:    ', $num, 'lightyellow','', false));
-            self::printResult('     Status:    ', $err->status ?? null, 'lightyellow','');    
-            
-            if (!empty($err->source) && !empty($err->source->parameter)){
-                self::printResult('     Parameter: ', $err->source->parameter, 'lightyellow');    
-            }
-            self::printResult('     Title:     ', $err->title ?? null, 'lightyellow');    
-            self::printResult('     Detail:    ', $err->detail ?? null, 'lightyellow');    
-
-            // separate errors
-            if (count($response->errors) > 1){
-                Console::log('   ---');
-            }
-        }
-        Console::log();           
-    }
-
-    /**
-     * Print a single error
-     * 
-     * @access protected
-     * @static
-     * @param string    $error      The error message
-     * 
-     * @return void
-     */
-    protected static function error(string $error)
-    {
-        if (self::isDefaultOuput()) {
-            // ✗
-            Console::log('  ' .   Console::text(' ERROR ','white', 'red'));
-            Console::log(
-                Console::text('   ✗', 'red') . 
-                Console::text(' Detail:    ', 'white') . 
-                Console::text($error, 'lightyellow') . 
-                Console::text('', 'white')
-            );    
-            Console::log();
-        }    
-    }
-    
-    /**
-     * Helper to validate a condition or exit with an error
-     * 
-     * @access protected
-     * @static
-     * @param bool      $condition      The condition to evaluate
-     * @param string    $message        Error message
-     * @param bool      $print          True to print error. Default is true
-     * 
-     * @return bool
-     */
-    protected static function validate(bool $condition, string $message, bool $print = true)
-    {
-        if ( !$condition ){
-            if ($print && self::isDefaultOuput()) {
-                Console::log();
-                self::error($message);
-                self::printFooter();
-            }
-            Program::exit(1);
-        }
-    }
-
-    /**
-     * Get numeric parameter and exit on error
-     * 
-     * @access protected
-     * @static
-     * @param array     $arguments
-     * @param string    $shortArg           The short argument name
-     * @param string    $longArg            The long argument name
-     * @param int       $defaultValue
-     * 
-     * @return int
-     */
-    protected static function getNumericParameter(array $arguments, string $shortArg, string $longArg, int $defaultValue)
-    {
-         if (self::inArguments($arguments,$shortArg, $longArg)){
-            $val = self::getArgumentValue($arguments,$shortArg, $longArg);
-
-            if (!is_numeric($val)){
-                self::error("Invalid parameter: $longArg must be a numeric value.");
-                self::printFooter();
-                Program::exit(1);
-            }
-            return intval($val);
-        }
-        return $defaultValue;
-    }
-
 }
